Clean up exposecontroller configuration (#8)
Configure exposecontroller to watch services only in current namespace
as its own serviceaccount, removing setup-time configuration for which
namespace will be watched. This is preparation for bundling the operator
for operator catalogs, so that yaml definitions do not contain any
"REPLACE_IMAGE" or "REPLACE_PROJECT" type variables to be run through
sed.