Replace exposecontroller with direct Routes
Remove exposecontroller in favour of directly generated OpenShift
Routes. This potentially removes compatibility with other Kubernetes
distributions, but allows the Operator to rely on the OpenShift
configured Router setup much more robustly without any user intervention
to configure the exposecontroller to match the Router.

This also adds the "bundle" subdirectory, which contains some Operator
catalog bundle definitions so that the Operator can be deployed to a
cluster by graphically installing it through the OpenShift Console, and
eventually can be published to OperatorHub so that the custom Operator
Source is not required.