Wrap Pod into Deployment object
This makes sense as an abstraction layer, even if the Deployment
replica management and other features are not necessarily useful at this
time. The Deployment is also the unit visible on the OpenShift 4
Topology view, so this change allows ContainerJFR resources to become
easily discoverable for Developer accounts within the namespace.

Related to #25