Grafana uses generated basic auth credentials
Generate a random password string and hardcoded username, and store
these in a Secret. Configure Grafana container to mount this Secret to
environment variables, causing the credentials to be used rather than
the insecure defaults. When configuring the Grafana instance's dashboard
and datasources, retrieve the Secret and use the credentials for HTTP
Basic request auth.

Related to #50
Related to #51